abstract station-popup to a new file
